pyproject.toml
[project]
name = "quantaquill-clone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "arxiv>=2.2.0",
    "fastapi>=0.116.1",
    "langchain>=0.3.26",
    "langchain-groq>=0.3.5",
    "markdown-pdf>=1.7",
    "matplotlib>=3.10.3",
    "networkx>=3.4.2",
    "pydantic>=2.11.7",
    "pypandoc>=1.15",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "scholarly>=1.7.11",
    "streamlit>=1.47.1",
    "uvicorn>=0.35.0",
]
